requirements.txt
numpy==1.24.1
flake8==6.0.0 
pandas==1.5.2
seaborn==0.12.2
scikit-learn==1.2.0
scikit-image==0.19.3
tqdm==4.64.1
opencv-python==4.7.0.68
pydantic==1.10.4
fastapi==0.89.0
uvicorn==0.20.0
Pillow==9.4.0
requests==2.28.1
python-multipart==0.0.5
Jinja2==3.1.2
jupyter==1.0.0
requests-toolbelt==0.10.1
awsebcli==3.20.3